refactor: Separate auto and manual load tests GitHub Actions
