--- .github/workflows/test-load-test.yml
@@ -1,37 +1,0 @@
-name: k6 Load Test (TEST)
-
-on: [workflow_dispatch, deployment_status]
-
-jobs:
-  k6:
-    if: |
-      (github.event.deployment_status.state == 'success' && github.ref_name == 'main') ||
-      github.event_name == 'workflow_dispatch'
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v1
-
-      - name: Testing GraphQL - Components
-        if: always()
-        uses: grafana/k6-action@v0.2.0
-        with:
-          filename: k6/graphql/test/components.js
-          cloud: true
-          token: ${{ secrets.K6_CLOUD_API_TOKEN }}
-
-      - name: Testing GraphQL - Observations
-        if: always()
-        uses: grafana/k6-action@v0.2.0
-        with:
-          filename: k6/graphql/test/observations.js
-          cloud: true
-          token: ${{ secrets.K6_CLOUD_API_TOKEN }}
-
-      - name: Testing HAR - Edition
-        if: always()
-        uses: grafana/k6-action@v0.2.0
-        with:
-          filename: k6/har/test/edition.js
-          cloud: true
-          token: ${{ secrets.K6_CLOUD_API_TOKEN }}
