--- k6/graphql/int/utils.js
@@ -1,1 +1,0 @@
-export const url = "https://int.visualize.admin.ch/api/graphql";
