--- k6/graphql/prod/utils.js
@@ -1,1 +1,0 @@
-export const url = "https://visualize.admin.ch/api/graphql";
