--- k6/graphql/test/utils.js
@@ -1,1 +1,0 @@
-export const url = "https://test.visualize.admin.ch/api/graphql";
